+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Change Log
 
 ## [master](https://github.com/arangodb/kube-arangodb/tree/master) (N/A)
-- Add graceful shutdown for coordinators
+- Add graceful shutdown for coordinators  Available since
+  ArangoDB versions: 3.7.12, 3.8.1, 3.9.0
 
 ## [1.2.3](https://github.com/arangodb/kube-arangodb/tree/1.2.3) (2021-09-24)
 - Update UBI Image to 8.4